+++ wasmbinding/test/custom_query_test.go
@@ -203,7 +203,6 @@ func TestQueryEstimateSwap(t *testing.T) {
 	amountIn := sdk.NewInt(10000)
 	query := bindings.OsmosisQuery{
 		EstimateSwap: &bindings.EstimateSwap{
-			Sender: reflect.String(),
 			First: bindings.Swap{
 				PoolId:   starPool,
 				DenomIn:  "uosmo",
@@ -232,7 +231,6 @@ func TestQueryEstimateSwap(t *testing.T) {
 	amountOut := sdk.NewInt(10000)
 	query = bindings.OsmosisQuery{
 		EstimateSwap: &bindings.EstimateSwap{
-			Sender: reflect.String(),
 			First: bindings.Swap{
 				PoolId:   starPool,
 				DenomIn:  "uosmo",